mysite/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.2
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
